Correcciones
tipos de datos, nombre de columnas y cantidad de caracteres, relaciones en entity (a excepcion de la relacion muchos a muchos, falta)

+++ ProyectoConMS/pedido-ms/src/main/java/cl/fullstack/pedido_ms/dto/DetallePedidoDTO.java
@@ -6,9 +6,9 @@ import lombok.Data;
 
 public class DetallePedidoDTO {
 
-    private Long pedidoId;
-    private Long productoId;
-    private Long idDetalle;
+    private int pedidoId;
+    private int productoId;
+    private int idDetalle;
     private int cantidad;
     private double precio;
