+++ ProyectoConMS/pruducto-ms/src/main/java/cl/fullstack/pruducto_ms/entity/ProveedorEntity.java
@@ -10,10 +10,10 @@ public class ProveedorEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "proveedor_seq")
-        //sirve para generar desde aca la seq de numeracion automatica, pasa que oracle no soporta generation type solo
+    //sirve para generar desde aca la seq de numeracion automatica, pasa que oracle no soporta generation type solo
     @SequenceGenerator(name = "proveedor_seq", sequenceName = "PROVEEDOR_SEQ", allocationSize = 1)
     @Column(name = "id")
-    private Long id;
+    private int id;
 
     @Column(name = "nombre", nullable = false, length = 20)
     private String nombre;
