+++ ProyectoConMS/pruducto-ms/src/main/java/cl/fullstack/pruducto_ms/repository/CategoriaRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 import cl.fullstack.pruducto_ms.entity.CategoriaEntity;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Long> {
+public interface CategoriaRepository extends JpaRepository<CategoriaEntity, Integer> {
 }
