+++ ProyectoConMS/pruducto-ms/src/main/java/cl/fullstack/pruducto_ms/service/IProveedorService.java
@@ -6,8 +6,8 @@ import cl.fullstack.pruducto_ms.dto.ProveedorDTO;
 
 public interface IProveedorService {
     List<ProveedorDTO> getAllProveedores();
-    ProveedorDTO getProveedorById(Long id);
+    ProveedorDTO getProveedorById(int id);
     ProveedorDTO createProveedor(ProveedorDTO proveedorDTO);
-    ProveedorDTO updateProveedor(Long id, ProveedorDTO proveedorDTO);
-    void deleteProveedor(Long id);
+    ProveedorDTO updateProveedor(int id, ProveedorDTO proveedorDTO);
+    void deleteProveedor(int id);
 }
